migrate test_dig to pytest
